+++ api/src/anonymization/referent.js
@@ -1,5 +1,5 @@
-const { generateRandomName, generateRandomEmail, generateNewPhoneNumber } = require("../utils/anonymise");
-const { anonymizeNonDeclaredFields } = require("../utils/anonymise-model-fields");
+const { generateRandomName, generateRandomEmail, generateNewPhoneNumber } = require("./utils/anonymise");
+const { anonymizeNonDeclaredFields } = require("./utils/anonymise-model-fields");
 
 function anonymize(itemToAnonymize) {
   const whitelist = [
